+++ src/apps/chifra/pkg/node/doc.go
@@ -1,2 +1,0 @@
-// Package node queries the node for various capabilities
-package node