+++ src/apps/chifra/pkg/rpcClient/is_archive.go
@@ -1,14 +1,14 @@
-package node
+package rpcClient
 
 import (
 	"path/filepath"
 
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/config"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/prefunds"
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/rpcClient"
 )
 
-func IsArchiveNode(chain string) bool {
+// IsNodeArchive returns true if the node is an archive node
+func IsNodeArchive(chain string) bool {
 	// TODO: from C++ code
 	// const CToml* config = getGlobalConfig("blockScrape");
 	// if (!config->getConfigBool("requires", "archive", true))
@@ -24,7 +24,7 @@ func IsArchiveNode(chain string) bool {
 		return false
 	}
 
-	bal, err := rpcClient.GetBalanceAt(chain, largest.Address, 0)
+	bal, err := GetBalanceAt(chain, largest.Address, 0)
 	if err != nil {
 		return false
 	}