+++ src/apps/chifra/pkg/rpcClient/is_tracing.go
@@ -1,8 +1,9 @@
-package node
+package rpcClient
 
 // TODO: BOGUS This needs to be implemented in a cross-chain, cross-client manner
 
-func IsTracingNode(testMode bool, chain string) bool {
+// IsNodeTracing returns true if the node is an archive node. Note currently always returns true.
+func IsNodeTracing(testMode bool, chain string) bool {
 	// TODO: We can test this with a unit test
 	if testMode && chain == "non-tracing" {
 		return false